+++ presets/ragengine/guardrails/output_guardrails.py
@@ -16,114 +16,20 @@ import re
 from dataclasses import dataclass, field
 from typing import Any
 
-import llm_guard.output_scanners as llm_guard_output_scanners
 import yaml
 from llm_guard import scan_output
-from llm_guard.input_scanners.ban_substrings import (
-    MatchType as BanSubstringsMatchType,
-)
-from llm_guard.input_scanners.regex import MatchType as RegexMatchType
 
 from ragengine import config
+from ragengine.guardrails.scanner_schemas import (
+    SCANNER_REGISTRY,
+    ParsedScannerConfig,
+)
 from ragengine.models import ChatCompletionResponse, get_message_content
 
 logger = logging.getLogger(__name__)
 
 DEFAULT_BLOCK_MESSAGE = "The model output was blocked by output guardrails."
 DEFAULT_ACTION_ON_HIT = "redact"
-
-
-# ===============================
-# Scanner config schemas
-#
-# Each schema describes the YAML shape AND knows how to build the
-# corresponding llm_guard scanner. To add a new scanner:
-#   1. Define a dataclass with from_dict() and build()
-#   2. Register it in SCANNER_REGISTRY below
-#
-# TODO: Once this scanner config surface stabilizes, move schema
-# validation to the admission webhook so that invalid policies are
-# rejected at apply-time instead of being silently skipped at runtime.
-#
-# TODO: Many llm_guard scanners (e.g. Toxicity, Bias, Language) do not
-# support redaction; pairing them with action=redact would be a no-op.
-# When adding such scanners, declare a per-schema `supports_redact` flag
-# and reject the (action=redact + non-redact scanner) combination here
-# in the schema layer, instead of trying to fix it at runtime.
-# ===============================
-
-
-@dataclass
-class BanSubstringsConfig:
-    substrings: list[str]
-    match_type: str = "word"
-    case_sensitive: bool = False
-    contains_all: bool = False
-
-    @classmethod
-    def from_dict(cls, raw: dict) -> "BanSubstringsConfig":
-        substrings = _coerce_string_list(raw.get("substrings"))
-        if not substrings:
-            raise ValueError(
-                "ban_substrings requires 'substrings' to be a non-empty list of strings"
-            )
-        return cls(
-            substrings=substrings,
-            match_type=str(raw.get("match_type", "word")).lower(),
-            case_sensitive=bool(raw.get("case_sensitive", False)),
-            contains_all=bool(raw.get("contains_all", False)),
-        )
-
-    def build(self, action_on_hit: str) -> Any:
-        return llm_guard_output_scanners.BanSubstrings(
-            substrings=list(self.substrings),
-            match_type=BanSubstringsMatchType[self.match_type.upper()],
-            case_sensitive=self.case_sensitive,
-            contains_all=self.contains_all,
-            redact=(action_on_hit == "redact"),
-        )
-
-
-@dataclass
-class RegexConfig:
-    patterns: list[str]
-    is_blocked: bool = True
-    match_type: str = "search"
-
-    @classmethod
-    def from_dict(cls, raw: dict) -> "RegexConfig":
-        patterns = _coerce_string_list(raw.get("patterns"))
-        if not patterns:
-            raise ValueError(
-                "regex requires 'patterns' to be a non-empty list of strings"
-            )
-        return cls(
-            patterns=patterns,
-            is_blocked=bool(raw.get("is_blocked", True)),
-            match_type=str(raw.get("match_type", "search")).lower(),
-        )
-
-    def build(self, action_on_hit: str) -> Any:
-        return llm_guard_output_scanners.Regex(
-            patterns=list(self.patterns),
-            is_blocked=self.is_blocked,
-            match_type=RegexMatchType[self.match_type.upper()],
-            redact=(action_on_hit == "redact"),
-        )
-
-
-SCANNER_REGISTRY: dict[str, type] = {
-    "ban_substrings": BanSubstringsConfig,
-    "regex": RegexConfig,
-}
-
-
-@dataclass
-class ParsedScannerConfig:
-    """A scanner config that has already passed schema validation."""
-
-    type: str
-    config: Any
 
 
 @dataclass
@@ -250,12 +156,6 @@ class OutputGuardrails:
                 prompt_parts.append(content)
 
         return "\n\n".join(prompt_parts)
-
-
-def _coerce_string_list(value: Any) -> list[str]:
-    if not isinstance(value, list):
-        return []
-    return [item for item in value if isinstance(item, str) and item]
 
 
 def _parse_policy_scanner_configs(
